Make native rejection second-class
An expanded version of this text should probably go into an RFC.
Basically, the only purpose of promise rejection seems to be
short-circuiting evaluation of Promise.then. However, that is properly a
job for the option and result types.

Combined with the difficulties of giving composable types to rejection,
it seems better to support rejection only enough that bindings authors
can easily convert rejections to some kind of explicit error handling,
the design of which we haven't yet settled on.

To that end, the previous Repromise API, which fully exposes both
fulfillment and rejection, is now in the second-class
Repromise.Rejectable module. Repromise.Rejectable is meant only for use
by bindings authors.

The base Repromise module now contains a type of promises that can only
become fulfilled.