Avoid passing executor to Repromise.new_
The original motivation for passing resolve and reject to a callback
called an "executor" seems to be detailed here:

  https://blog.domenic.me/the-revealing-constructor-pattern/

However, in Reason, we can achieve all the same goals by returning
pairs of promises and resolver functions instead (or triples, in
Repromise.Rejectable). Since this is less awkward and more idiomatic
for Reason, this commit changes new_ to take () and return a pair.